Fix coverage gap, refactor link reference handling
